refactor: inline legacy slack shared handlers into slack-org module
Move fetchConversationContexts, getWorkspaceAgent, resolveSessionCompose,
and enrichMessageContent directly into slack-org/handlers/shared.ts,
eliminating the re-export indirection. Delete the now-unused legacy
slack/handlers/shared.ts, run-agent.ts, and its test.
